Rename devices module to device
