+++ src/arch/x86/device.rs
@@ -1,4 +1,4 @@
-use devices::{pic_8259, ps2_controller_8042, uart_16550};
+use device::{pic_8259, ps2_controller_8042, uart_16550};
 
 pub fn init() {
     pic_8259::init();
